+++ src/imio/project/pst/content/action.py
@@ -2,8 +2,6 @@
 from Acquisition import aq_inner
 from collective.contact.plonegroup.utils import organizations_with_suffixes
 from collective.task.interfaces import ITaskContent
-from collective.z3cform.chosen.widget import AjaxChosenFieldWidget
-from collective.z3cform.chosen.widget import AjaxChosenMultiFieldWidget
 from collective.z3cform.datagridfield import DataGridFieldFactory
 # from dexterity.localrolesfield.field import LocalRoleField
 from collective.z3cform.datagridfield import DictRow
@@ -78,14 +76,12 @@ class IPSTAction(IProject):
         ),
         required=False,
     )
-    form.widget('representative_responsible', AjaxChosenMultiFieldWidget, populate_select=True)
 
     responsible = schema.Choice(
         title=_(u'Responsible'),
         vocabulary=u'imio.project.pst.ActionEditorsVocabulary',
         required=False,
     )
-    form.widget('responsible', AjaxChosenFieldWidget, populate_select=True)
 
     health_indicator = schema.Choice(
         title=_(u'Health indicator'),